store/wizard: extract types to dedicated module
Rename wizardState to WizardState for consistency. Having types
in a standalone file avoids circular dependencies when submodules
need to reference the parent state type without importing the
full slice.
