+++ src/store/slices/wizard/index.ts
@@ -1,4 +1,5 @@
 export * from './slice';
+export * from './types';
 
 // submodules
 export * from './filesystem';
