+++ src/store/slices/wizard/tests/targets.test.ts
@@ -15,7 +15,7 @@ import {
   reinitializeAzure,
   reinitializeGcp,
   wizardReducer,
-  type wizardState,
+  type WizardState,
 } from '@/store/slices/wizard';
 
 describe('target environment reducers', () => {
@@ -42,7 +42,7 @@ describe('target environment reducers', () => {
       });
 
       it('should clear source ID with undefined', () => {
-        const stateWithSource: wizardState = {
+        const stateWithSource: WizardState = {
           ...initialState,
           cloudProviders: {
             ...initialState.cloudProviders,
@@ -75,7 +75,7 @@ describe('target environment reducers', () => {
 
     describe('reinitializeAws', () => {
       it('should reset all AWS fields to defaults', () => {
-        const modifiedState: wizardState = {
+        const modifiedState: WizardState = {
           ...initialState,
           cloudProviders: {
             ...initialState.cloudProviders,
@@ -98,7 +98,7 @@ describe('target environment reducers', () => {
       });
 
       it('should not affect other state', () => {
-        const modifiedState: wizardState = {
+        const modifiedState: WizardState = {
           ...initialState,
           cloudProviders: {
             ...initialState.cloudProviders,
@@ -181,7 +181,7 @@ describe('target environment reducers', () => {
 
     describe('reinitializeAzure', () => {
       it('should reset all Azure fields to undefined', () => {
-        const modifiedState: wizardState = {
+        const modifiedState: WizardState = {
           ...initialState,
           cloudProviders: {
             ...initialState.cloudProviders,
@@ -202,7 +202,7 @@ describe('target environment reducers', () => {
       });
 
       it('should preserve hyperVGeneration', () => {
-        const modifiedState: wizardState = {
+        const modifiedState: WizardState = {
           ...initialState,
           cloudProviders: {
             ...initialState.cloudProviders,
@@ -274,7 +274,7 @@ describe('target environment reducers', () => {
 
     describe('reinitializeGcp', () => {
       it('should reset all GCP fields to defaults', () => {
-        const modifiedState: wizardState = {
+        const modifiedState: WizardState = {
           ...initialState,
           cloudProviders: {
             ...initialState.cloudProviders,
